Make descriptor to ProtoData AST conversion public
